FEBE/backend/requirements.txt
fastapi
uvicorn
scikit-learn
pydantic
python-jose
passlib[bcrypt]
h5py
tensorflow